Changed everything again.
Turned out that there was no way to make a SOCK_HASH map the way I wanted to
and no great benefit either. I went for the simplest, if not ideal: a
sock_filter program with an aribtrary allowed port list (of length 1, right
now). That seems to work (at least it loads).