Reduced to a one function library.
Removed all the socket opening aspect from the code. This is now
just a library to load a port filter ebpf program/map.

The rest was moved to the test, in the form of opening a socket and attaching
the program to it. Which the afpacket library does nicely for us.