chore: tools dependencies cleanup (#4760)
1. `buf` is run with `go run`.
2. `golangci-lint` is run with `go run`.
3. `impi` is replaced with `goimports` linter in `golangci-lint`.
4. Custom `ineffassign` is removed (there's already `ineffassign` linter
in `golangci-lint`).
5. Tools dependencies that were required for `impi` and custom
`ineffassign` are removed.
6. Both `impi` and custom `ineffassign` were used in a wrapper for
`go_library`, and there's no extra functionality in that wrapper.
Therefore this wrapper is deleted and the default `go_library` rule is
used, causing a myriad of changes in `BUILD.bazel` files.
7. Both `impi` and custom `ineffassign` were used in a wrapper for
`go_test` too. There is extra functionality in that wrapper, so it's
still there. It's not related to the linters anymore, so this wrapper is
moved, also causing changes in `BUILD.bazel` files.