doc: typos (#4764)
This pull request includes minor documentation updates to fix
typographical errors and improve clarity in SCION's technical
documentation.

Typographical corrections:

*
[`doc/overview.rst`](diffhunk://#diff-6cc65603a3c60d4383d1e05dd6388e73773844c2f8e981f30d1ed116a7177e1bL14-R14):
Corrected a grammatical error by changing "focuses on providing" to
"focus on providing" in the description of SCION's aspirations for
inter-AS routing.
*
[`doc/sig.rst`](diffhunk://#diff-c28ddb50c1be3aaf4c126dd3dbc0ae78e427b6d23017c717cd570bd9783ba544L88-R88):
Fixed a typographical error by correcting "out-ot-order" to
"out-of-order" in the explanation of SIG frame handling.