tests: let graphupdater generate multiple graphs (#4766)
There's only one graph that is available for tests. It's generated with
`graphupdater` tool which doesn't create the second graph properly
because the same variable names might be used in the generated code.
This PR fixes it, and adds one more graph that is an example of how the
code with multiple graphs will look like. This graph will be used in the
future for https://github.com/scionproto/scion/issues/4699
implementation.