tests: generate topology.json files using the graph info (#4767)
This PR adds an ability to use `topogen` to generate `topology.json`
files that are used in tests. By default, all interface IDs generated by
the tool are random, while the `graph` package defines deterministic
interface IDs. The code here patches interface IDs in the
`topology.json` files so they are the same as defined in the `graph`
package, and adds Bazel rules to generate the files that can be used in
the tests. The test data that is added here serves only demonstration
purposes for now, but will be later used for
https://github.com/scionproto/scion/issues/4699.

---------

Co-authored-by: Roman Sharkov <roman.scharkov@gmail.com>