cppki: correct struct tag asn1 (#4778)
I don't know how the asn1 encoder/decoder behaves but I assume that they
just fall back to case-insensitive matching by field name so nothing
broke despite the tag being wrong. That is assuming that it behaves
similarly to how `encoding/json` maps fields. So this is most likely not
a bug fix but a minor typo fix.