Improve dropped packets metrics and give one more cpu to the traffic generator.
The issue with the traffic genertaor is that it struggles to saturate the
router, but if we make it too fast (e.g. giving it a performance core), it
seems to outcompete the router for resources. I am guessing that using veth
creates some contention in the kernel. Not sure.