test: use capsh at runtime instead of setcap at buildtime (#4802)
Bazel doesn't deal properly with extended file attributes. The original
plan was to produce the capabilities enabled binary by copying the
normal one and applying setcap to the result, but in some cases the copy
can be left without the attributes and bazel considering the goal
reached anyway. To work around that, wrapped the test in a script that
executes the normal binary via capsh.