path-combinator: fix path mtu when segment is shortcut (#4805)
Fixes a bug in the path construction logic.

If only use a suffix of a segment is used (shortcut path) the ingress
MTU of the last ASentry that is actually used should not be included in
the path MTU calculation.

Example:
In the topology below, if there is a segment from 130 to 112.
If the segment is used to build a path from 112 to 111, the resulting
MTU should be 3 (previously it would have been 2).
```
    MTU
┌───┐  
│130| 1
└─┬─┘  
  │ 2  
┌─▼─┐  
│111| 3
└─┬─┘  
  │ 4  
┌─▼─┐  
│112| 5
└───┘  
```