beaconing: fix diversity calculation
Currently the beacon diversity calculation was wrong. We only considered
ingress interface in construction direction. However the first ASEntry
will always have a zero ingress interface. Furthermore with this
approach we did not really consider the last ingress interface, which
was the ingress interface into the AS, which is not yet represented by
an ASEntry (as the termination/extension only happens after selection).

By using the egress interfaces interfaces instead we solve both
problems.